requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
pandas == 1.1.2
xlrd
loguru
flask-cors